pyproject.toml
[project]
name = "xgenpcb"
version = "0.1.0"
description = "Autonomous AI Agent for PCB Design & Synthesis"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "rich>=13.7.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.1",
    "websockets>=12.0",
]

[project.scripts]
xgenpcb = "xgenpcb.cli:main_wrapper"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["xgenpcb"]

